Implementation of Details portlet
- In Visualization portlet we capture the onClick event,
i.e. when the user clicks on an entity and this propagates in
Details portlet where it updates the panel with useful data of the
selected entity.
- We don't use the dispatcher in Details portlet, i.e.
send a request and receive a response, as we do it - for the time being -
client-side, e.g. in Visualization portlet we already have a json string of
converted ontology, so we use JSONParser (client-based only) to transform
the string to a JSONObject and then we render a Panel object.
- Details portlet which listens for onClick events, gets the change and renders a
new panel.
- This approach might not be efficient; it is under development.
- Another possible solution (yet to be determined) if we should follow the JSNI approach
which is followed in Visualization portlet, i.e. call a Javascript function (in webvowl-app.js)
to render the details panel (as an HTML widget).